Perf judge builder with `PackageAgent` trait
